Allow #include <gsl/...> without repeating gsl
Headers are named inconsistently: algorithm, assert, byte and util
headers are prefixed with gsl_. This is to avoid name clash with
standard headers for people including directly from the gsl directory.

However, for people including from the parent directory of gsl, in
addition to looking inconsistent, this also looks redundant:
#include <gsl/gsl_assert>

This adds algorithm, assert, byte and util headers that just forward to
gsl_ headers. It is then possible to include like this:
#include <gsl/assert>

Fixes #71