Add throw exception variant for GSL_UNENFORCED_ON_CONTRACT_VIOLATION
There are three configuration options for when conditions are violated.
When GSL_UNENFORCED_ON_CONTRACT_VIOLATION is selected, nothing
should happen, but throw_exception()'s behavior is not consistent
with this.

The ifdefs for the three config options (throw, terminate, unenforced)
have been reordered to be consistent with the order they appear in
the documentaton at the start of the header as well as the definition
of GSL_CONTRACT_CHECK right below these changes

#818