Ignore ABI compatibility warnings with old GCC
We can see "note: the layout of aggregates containing vectors with 4-byte
alignment has changed in GCC 5" on the standard <regex> header when compiling
for PowerPC 64-bit LE target. This results in a test failure due to output to
stderr. The warning looks useless since GSL does not support so old GCC 5.